+++ src/pptx/oxml/__init__.py
@@ -397,6 +397,7 @@ from pptx.oxml.slide import (  # noqa: E402
     CT_Background,
     CT_BackgroundProperties,
     CT_CommonSlideData,
+    CT_HeaderFooter,
     CT_NotesMaster,
     CT_NotesSlide,
     CT_Slide,
@@ -413,6 +414,7 @@ register_element_cls("p:bg", CT_Background)
 register_element_cls("p:bgPr", CT_BackgroundProperties)
 register_element_cls("p:childTnLst", CT_TimeNodeList)
 register_element_cls("p:cSld", CT_CommonSlideData)
+register_element_cls("p:hf", CT_HeaderFooter)
 register_element_cls("p:notes", CT_NotesSlide)
 register_element_cls("p:notesMaster", CT_NotesMaster)
 register_element_cls("p:sld", CT_Slide)